+++ crates/ai/src/providers/open_ai/mod.rs
@@ -1,4 +1,7 @@
 pub mod completion;
+pub mod embedding;
 pub mod model;
+
 pub use completion::*;
+pub use embedding::*;
 pub use model::OpenAILanguageModel;
